+++ packages/common/src/core/schemas/types.ts
@@ -11,10 +11,16 @@ import { TemplateEditorTypes } from "../../templates/_internal/TemplateSchema";
 import { GroupEditorTypes } from "../../groups/_internal/GroupSchema";
 import { InitiativeTemplateEditorTypes } from "../../initiative-templates/_internal/InitiativeTemplateSchema";
 import { SurveyEditorTypes } from "../../surveys/_internal/SurveySchema";
+import {
+  CardEditorOptions,
+  EntityEditorOptions,
+} from "./internal/EditorOptions";
+import { IArcGISContext } from "../../ArcGISContext";
 
 export interface IEditorConfig {
   schema: IConfigurationSchema;
   uiSchema: IUiSchema;
+  defaults?: IConfigurationValues;
 }
 
 /**
@@ -68,6 +74,50 @@ export const validEditorTypes = [
   ...validEntityEditorTypes,
   ...validCardEditorTypes,
 ] as const;
+
+/**
+ * An editor's module when dynamically imported depending on the EditorType. This
+ * will always have a buildUiSchema function, and sometimes it will have a
+ * buildDefaults function to override default values in the editor.
+ */
+export type IEditorModuleType = IEntityEditorModuleType | ICardEditorModuleType;
+
+/**
+ * An entity editor's module when dynamically imported depending on the EditorType. This
+ * will always have a buildUiSchema function, and sometimes it will have a
+ * buildDefaults function to override default values in the editor.
+ */
+export interface IEntityEditorModuleType {
+  buildUiSchema: (
+    i18nScope: string,
+    options: EntityEditorOptions,
+    context: IArcGISContext
+  ) => Promise<IUiSchema>;
+  buildDefaults?: (
+    i18nScope: string,
+    options: EntityEditorOptions,
+    context: IArcGISContext
+  ) => Promise<IConfigurationValues>;
+}
+
+/**
+ * A card editor's module when dynamically imported depending on the EditorType. This
+ * will always have a buildUiSchema function, and sometimes it will have a
+ * buildDefaults function to override default values in the editor.
+ */
+export interface ICardEditorModuleType {
+  buildUiSchema: (
+    i18nScope: string,
+    config: CardEditorOptions,
+    context: IArcGISContext
+  ) => Promise<IUiSchema>;
+
+  buildDefaults?: (
+    i18nScope: string,
+    options: CardEditorOptions,
+    context: IArcGISContext
+  ) => Promise<IConfigurationValues>;
+}
 
 export enum UiSchemaRuleEffects {
   SHOW = "SHOW",